updated restaurant code to almost work with simcity market implementation

+++ src/city/interfaces/RestaurantChungCustomer.java
@@ -2,7 +2,6 @@ package city.interfaces;
 
 import utilities.RestaurantChungMenu;
 import city.animations.RestaurantChungCustomerAnimation;
-import city.animations.interfaces.RestaurantChungAnimatedCustomer;
 
 /**
  * A sample Customer interface built to unit test a CashierAgent.
@@ -23,11 +22,11 @@ public interface RestaurantChungCustomer extends RoleInterface {
 	public abstract void msgOutOfItem(String choice, RestaurantChungMenu menu);
 	public abstract void msgHereIsYourFood();
 	public abstract void msgSelfDoneEating();
-	public abstract void msgHereIsCheck(double price);
+	public abstract void msgHereIsCheck(int price);
 	public abstract void msgAnimationAtCashier();
-	public abstract void msgHereIsChange(double change);
+	public abstract void msgHereIsChange(int change);
 	public abstract void msgAnimationFinishedLeaveRestaurant();
-	public abstract void msgKickingYouOutAfterPaying(double debt);
+	public abstract void msgKickingYouOutAfterPaying(int debt);
 	public abstract void setGui(RestaurantChungCustomerAnimation g);
 	public abstract void setHost(RestaurantChungHost host);
 	public abstract void setCashier(RestaurantChungCashier cashier);
@@ -35,25 +34,5 @@ public interface RestaurantChungCustomer extends RoleInterface {
 	public abstract RestaurantChungCustomerAnimation getGui();
 	public abstract String getState();
 	public abstract String getOrder();
-	/**
-	 * @param total The cost according to the cashier
-	 *
-	 * Sent by the cashier prompting the customer's money after the customer has approached the cashier.
-	 */
-//	public abstract void HereIsYourTotal(double total);
-
-	/**
-	 * @param total change (if any) due to the customer
-	 *
-	 * Sent by the cashier to end the transaction between him and the customer. total will be >= 0 .
-	 */
-//	public abstract void HereIsYourChange(double total);
-
-
-	/**
-	 * @param remaining_cost how much money is owed
-	 * Sent by the cashier if the customer does not pay enough for the bill (in lieu of sending {@link #HereIsYourChange(double)}
-	 */
-//	public abstract void YouOweUs(double remaining_cost);
 
 }
