+++ src/city/roles/RestaurantChungWaiterRevolvingStandRole.java
@@ -30,7 +30,7 @@ public class RestaurantChungWaiterRevolvingStandRole extends RestaurantChungWait
 //			String choice;
 //			OrderStatus os;
 //		}
-//		double bill;
+//		int bill;
 //		
 //		public WCustomer(Customer customer, int table, CustomerState state) {
 //			c = customer;
@@ -146,7 +146,7 @@ public class RestaurantChungWaiterRevolvingStandRole extends RestaurantChungWait
 
 //	Cashier
 //	---------------------------------------------------------------
-	public void msgHereIsBill(RestaurantChungCustomer c, double price) {
+	public void msgHereIsBill(RestaurantChungCustomer c, int price) {
 		print("Waiter received msgHereIsBill");
 		WCustomer wc = findCustomer(c);
 		wc.bill = price;
