+++ src/city/tests/mock/MockRestaurantChungWaiterRevolvingStand.java
@@ -96,7 +96,7 @@ public class MockRestaurantChungWaiterRevolvingStand extends MockRole implements
 	}
 
 	@Override
-	public void msgHereIsBill(RestaurantChungCustomer c, double price) {
+	public void msgHereIsBill(RestaurantChungCustomer c, int price) {
 		log.add(new LoggedEvent("Waiter received msgHereIsBill from cashier. The bill is for " + price));
 		System.out.println("Waiter received msgHereIsBill from cashier. The bill is for " + price);			
 	}
